Automated_AI_Instagram/requirements.txt
# Requirements for Automated_AI_Instagram
pillow
requests
openai
pytest
pytest-mock
